+++ src/clj_kondo/impl/analysis.clj
@@ -3,6 +3,7 @@
   {:no-doc true}
   (:refer-clojure :exclude [ns-name])
   (:require
+   [clj-kondo.impl.overrides :as overrides]
    [clj-kondo.impl.utils :refer [assoc-some export-ns-sym select-some]]))
 
 (defn select-context [selector ctx]
@@ -53,7 +54,16 @@
                                     :protocol-ns :protocol-name
                                     :imported-ns
                                     :name-row :name-col :name-end-col :name-end-row
-                                    :arglist-strs :end-row :end-col])]
+                                    :arglist-strs :end-row :end-col])
+          overrides (overrides/overrides (case base-lang
+                                           :clj '{:clj {:defs {clojure.core {}}}}
+                                           :cljs '{:cljs {:defs {cljs.core {}}}}
+                                           :cljc '{:cljc {:defs {cljs.core {}}}}))
+          attr-overrides (get-in overrides (case base-lang
+                                             :clj `[:clj :defs clojure.core ~nom]
+                                             :cljs `[:cljs :defs cljs.core ~nom]
+                                             :cljc `[:cljc :defs cljs.core ~lang ~nom]))
+          attrs (merge attrs attr-overrides)]
       (swap! analysis update :var-definitions conj
              (assoc-some
               (cond-> (merge {:filename filename